feat(screens): basic login, email + password
